+++ auth/bin/coverage3
@@ -1,0 +1,8 @@
+#!/home/james/Tooling-Project/auth/bin/python3
+# -*- coding: utf-8 -*-
+import re
+import sys
+from coverage.cmdline import main
+if __name__ == '__main__':
+    sys.argv[0] = re.sub(r'(-script\.pyw|\.exe)?$', '', sys.argv[0])
+    sys.exit(main())
